--- app/controllers/application_controller.rb
@@ -1,4 +1,0 @@
-module MountainView
-  class ApplicationController < ActionController::Base
-  end
-end
