+++ app/controllers/mountain_view/styleguide_controller.rb
@@ -1,0 +1,9 @@
+module MountainView
+  class StyleguideController < ::ApplicationController
+    layout "mountain_view"
+
+    def show
+      @component = params[:id]
+    end
+  end
+end
