--- app/controllers/styleguide_controller.rb
@@ -1,9 +1,0 @@
-class StyleguideController < ApplicationController
-
-  layout "mountain_view"
-
-  def show
-    @component = params[:id]
-  end
-
-end
