+++ app/views/mountain_view/styleguide/index.html.erb
