+++ app/views/mountain_view/styleguide/show.html.erb
